adding logs for subscription test
