+++ .gitignore
@@ -17,6 +17,7 @@ packages/**/chain-setup/**/db-data
 # Logs
 logs
 *.log
+.parcel-cache
 
 # Diagnostic reports (https://nodejs.org/api/report.html)
 report.[0-9]*.[0-9]*.[0-9]*.[0-9]*.json
