+++ package.json
@@ -54,6 +54,7 @@
     "eslint-plugin-prettier": "^5.1.3",
     "husky": "^9.0.11",
     "path": "^0.12.7",
+    "string_decoder": "^1.3.0",
     "ts-node": "^10.9.2",
     "turbo": "^2.1.3",
     "typescript": "^5.5.3",
