+++ packages/chess-contracts/src/index.ts
@@ -4,6 +4,7 @@ export {
   Payment,
   WinnerTxWrapper,
   NotEnoughFundError,
+  signRedeemTx,
 } from './chess-contract.js'
 export { User, UserHelper } from './user.js'
 export { Chess } from './chess.js'
